+++ tests/testthat/test-read.R
@@ -398,88 +398,6 @@ test_that("Make cell bounding boxes when segmentation is absent", {
     unlink("vizgen_test", recursive = TRUE)
 })
 
-# Format transcript spots===================
-test_that("Read MERFISH transcript spots into rowGeometries", {
-    dir_use <- VizgenOutput("hdf5", file_path = "test_spots")
-    sfe <- readVizgen(dir_use, z = 3L, image = "PolyT", add_molecules = TRUE)
-    expect_equal(rowGeometryNames(sfe), "txSpots")
-    rg <- txSpots(sfe)
-    expect_equal(as.character(st_geometry_type(rg, FALSE)), "MULTIPOINT")
-    # Check that the spots are flipped and aligned with the image
-    img <- imgRaster(getImg(sfe))
-    v <- terra::extract(img, rg)
-    expect_true(sum(v$mosaic_PolyT_z3 < 30, na.rm = TRUE) < 10)
-    expect_false(anyNA(v$mosaic_PolyT_z3))
-    unlink("test_spots", recursive = TRUE)
-})
-
-test_that("Format MERFISH transcript spots for colGeometries", {
-    dir_use <- VizgenOutput("hdf5", file_path = "cg_vizgen")
-    expect_error(formatTxSpots(file.path(dir_use, "detected_transcripts.csv"),
-                               dest = "colGeometry"),
-                 "file_out must be specified")
-    expect_error(formatTxSpots(file.path(dir_use, "detected_transcripts.csv"),
-                               dest = "colGeometry", file_out = "vizgen_cellbound/tx_in_cells"),
-                 "Column indicating cell ID not found.")
-
-    df <- read.csv(file.path(dir_use, "detected_transcripts.csv"),
-                   header = TRUE)
-    sfe <- readVizgen(dir_use, z = 3L, image = "PolyT")
-    df$cell_id <- sample(colnames(sfe), nrow(df), replace = TRUE)
-    rownames(df) <- df$X
-    df$X <- NULL
-    write.csv(df, file = file.path(dir_use, "detected_transcripts.csv"),
-              row.names = TRUE, quote = FALSE)
-
-    # First run
-    cg <- formatTxSpots(file.path(dir_use, "detected_transcripts.csv"),
-                        dest = "colGeometry",
-                        file_out = file.path(dir_use, "tx_in_cells"),
-                        z = 3L)
-    dir_check <- file.path(dir_use, "tx_in_cells")
-    expect_equal(cg, dir_check)
-    expect_true(dir.exists(dir_check))
-    fns_expect <- paste0(unique(df$gene[df$global_z == 3L]), "_spots.parquet")
-    fns <- list.files(dir_check)
-    expect_setequal(fns, fns_expect)
-
-    # Check contents
-    fn <- file.path(dir_check, fns_expect[1])
-    g <- st_read(fn)
-    expect_equal(as.character(st_geometry_type(g, FALSE)), "MULTIPOINT")
-
-    # Second run
-    time_note <- Sys.time() # Check the files weren't written again
-    cg <- formatTxSpots(file.path(dir_use, "detected_transcripts.csv"),
-                        dest = "colGeometry",
-                        file_out = file.path(dir_use, "tx_in_cells"))
-    expect_equal(cg, normalizePath(dir_check))
-    time_check <- file.info(fn)$mtime
-    expect_true(time_note > time_check)
-
-    unlink("cg_vizgen", recursive = TRUE)
-})
-
-test_that("Error messages in formatTxSpots", {
-    dir_use <- VizgenOutput("hdf5", file_path = "vizgen_test")
-    file <- file.path(dir_use, "detected_transcripts.csv")
-    expect_error(formatTxSpots(file, dest = "colGeometry", file_out = NULL),
-                 "file_out must be specified")
-    file2 <- system.file("extdata/sfe_visium.rds", package = "SpatialFeatureExperiment")
-    expect_error(formatTxSpots(file2),
-                 "The file must be one of csv")
-    expect_error(formatTxSpots(file, z = "foo"),
-                 "z must either be numeric")
-    expect_error(formatTxSpots(file, spatialCoordsNames = c("foo", "bar")),
-                 "foo, bar not found")
-    expect_error(formatTxSpots(file, z = 8L),
-                 "z plane\\(s\\) specified not found")
-    expect_error(formatTxSpots(file, dest = "colGeometry", cell_col = "foo",
-                               file_out = "bar"),
-                 "Column indicating cell ID not found")
-    unlink("cg_vizgen", recursive = TRUE)
-})
-
 # Read CosMX===================
 test_that("readCosMX, not reading transcript spots", {
     dir_use <- CosMXOutput(file_path = "cosmx_test")
@@ -598,59 +516,6 @@ test_that("readCosMX, split z, split cell compartments", {
     unlink("cosmx_test", recursive = TRUE)
 })
 
-test_that("Format CosMX spots for colGeometry, multiple z-planes", {
-    dir_use <- CosMXOutput(file_path = "cosmx_test")
-
-    cg <- formatTxSpots(file.path(dir_use, "Run5642_S3_Quarter_tx_file.csv"),
-                        dest = "colGeometry", z = "all", z_option = "split",
-                        cell_col = c("cell_ID", "fov"),
-                        gene_col = "target", not_in_cell_id = "0",
-                        spatialCoordsNames = c("x_global_px", "y_global_px", "z"),
-                        file_out = file.path(dir_use, "tx_spots"))
-    expect_equal(cg, file.path(dir_use, "tx_spots"))
-    # Oh, great, there's Bex1/2, illegal file name. No wonder people don't like CosMX
-    df <- data.table::fread(file.path(dir_use, "Run5642_S3_Quarter_exprMat_file.csv"))
-    genes <- names(df)[-c(1:2)]
-    combs <- expand.grid(gene = genes, z = 0:1, stringsAsFactors = FALSE)
-    fns_expect <- paste0(gsub("/", ".", combs$gene), "_spots_z", combs$z, ".parquet")
-    fns <- list.files(file.path(dir_use, "tx_spots"))
-    # Not all genes have spots in this downsampled toy dataset
-    expect_true(all(fns %in% fns_expect))
-
-    fn <- file.path(dir_use, "tx_spots", fns[1])
-    g <- st_read(fn)
-    expect_equal(st_geometry_type(g, FALSE) |> as.character(), "MULTIPOINT")
-    unlink(dir_use, recursive = TRUE)
-})
-
-test_that("addTxSpots add a subset of spots", {
-    skip()
-    skip_if_not(gdalParquetAvailable())
-    fn <- XeniumOutput("v2", file_path = "xenium_test")
-    sfe <- readXenium(fn)
-    tx_fn <- formatTxSpots(file.path(fn, "transcripts.parquet"),
-                           gene_col = "feature_name",
-                           spatialCoordsNames = c("x_location", "y_location", "z_location"),
-                           z_option = "3d", file_out = "txSpots.parquet",
-                           return = FALSE)
-    expect_error({
-        tx <- formatTxSpots(file.path(fn, "transcripts.parquet"),
-                            gene_col = "feature_name",
-                            gene_select = rownames(sfe)[1:5],
-                            return = FALSE)
-    }, "file_out for the reformatted transcript spots must be specified")
-    # TODO: New function to selectively read genes from existing GeoParquet file
-    # Shouldn't require the original file name. Also add the swap_rownames
-    # argument.
-    sfe <- addTxSpots(sfe, gene_col = "feature_name",
-                      gene_select = rowData(sfe)$Symbol[1:5],
-                      file_out = "txSpots.parquet")
-})
-
-test_that("add subset of spots, multiple files in tx spots output", {
-
-})
-
 unlink("cosmx_test", recursive = TRUE)
 unlink("vizgen_test", recursive = TRUE)
 
@@ -666,13 +531,13 @@ test_that("readXenium, XOA v1", {
     expect_s4_class(sfe, "SpatialFeatureExperiment")
     expect_equal(colGeometryNames(sfe), c("centroids", "cellSeg", "nucSeg"))
     expect_equal(rowGeometryNames(sfe), "txSpots")
-    expect_equal(as.character(st_geometry_type(centroids(sfe), FALSE)), "POINT")
+    expect_equal(as.character(st_geometry_type(SpatialFeatureExperiment::centroids(sfe), FALSE)), "POINT")
     expect_equal(as.character(st_geometry_type(cellSeg(sfe), FALSE)), "POLYGON")
     expect_equal(as.character(st_geometry_type(txSpots(sfe), FALSE)), "MULTIPOINT")
     expect_equal(imageIDs(sfe), c("morphology_focus", "morphology_mip"))
     expect_s4_class(getImg(sfe, image_id = "morphology_focus"), "BioFormatsImage")
     expect_s4_class(getImg(sfe, image_id = "morphology_mip"), "BioFormatsImage")
-    expect_equal(dim(getImg(sfe, image_id = "morphology_focus"))["C"], 1L)
+    expect_equal(dim(getImg(sfe, image_id = "morphology_focus"))[["C"]], 1L)
 
     # That things are aligned
     bbox_rg <- st_bbox(txSpots(sfe)) |> st_as_sfc()
@@ -715,7 +580,7 @@ test_that("readXenium XOA v1 when only cell but not nuclei segmentation is avail
     file.remove(list.files(fn, "cell_boundaries\\.*", full.names = TRUE))
     expect_warning(sfe <- readXenium(fn), 'No `cell_boundaries` file is available')
     expect_equal(colGeometryNames(sfe), c("centroids", "nucSeg"))
-    expect_warnings(sfe2 <- readXenium(fn, segmentations = "cell"),
+    expect_warning(sfe2 <- readXenium(fn, segmentations = "cell"),
                     'No `cell_boundaries` file is available')
     expect_equal(colGeometryNames(sfe2), "centroids")
     unlink("xenium_test", recursive = TRUE)